% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check missing values.R
\name{check_missing_values}
\alias{check_missing_values}
\title{Check missing values function}
\usage{
check_missing_values(df, columns)
}
\arguments{
\item{df}{input dataset}

\item{columns}{a list of columns to check}
}
\value{
a dataframe with number of missing values for each column
}
\description{
This function help check missing values in certain columns in a dataset.
}
\author{
Xiyue Li
}
\keyword{Missing}
\keyword{values}
